refactor: Upgrade to model-2.0.62 - letter_sound_correspondence
ref #1627
