+++ src/main/java/ai/elimu/web/content/letter_sound/LetterSoundCorrespondenceCsvExportController.java
@@ -1,4 +1,4 @@
-package ai.elimu.web.content.letter_sound_correspondence;
+package ai.elimu.web.content.letter_sound;
 
 import ai.elimu.model.content.Sound;
 import ai.elimu.model.content.Letter;
