+++ src/main/java/ai/elimu/web/content/letter_sound/LetterSoundCorrespondencePeerReviewsController.java
@@ -1,4 +1,4 @@
-package ai.elimu.web.content.letter_sound_correspondence;
+package ai.elimu.web.content.letter_sound;
 
 import ai.elimu.dao.EmojiDao;
 import ai.elimu.dao.LetterSoundContributionEventDao;
